refactor 02
